[TRANSLATION] wallet/liana - IT (#1624)
* Create 04.webp

04.webp image no text

* Delete courses/biz101/assets/notext directory

* Update 022_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 023_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 024_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update 025_it.yml

Replaced Bitcoin with bitcoin as it indicates value as coin and not as protocol.

* Update Liana_it.md

First global review to be reviewed.

* Update Liana_it.md

First global review to be reviewed.

* Update it.md

* Update it.md

* Update it.md

* Update it.md

* Update it.md

* Update Wallet_Liana_it.md

I have arranged all the formatting making it equal to that of the reference translation in English

* Changing portafoglio into wallet

---------

Co-authored-by: MarJJ <mariannacon@protonmail.com>
Co-authored-by: asi0 <90136504+Asi0Flammeus@users.noreply.github.com>