[PROOFREADING] - HIS205 - EN (#3724)
* Update en.md - first review

* Update en.md - Reviewed until line 164

* Update en.md reviewed until line 291

* Update en.md - reviewed until line 339

* Update en.md - reviewed until line 482

* Update en.md - line 499

* Update en.md - reviwed until line 618

* Update en.md - reviewed until line 636

* Update en.md - reviewed until line 716

* Update en.md - line 755

* Update en.md line 917

* Update en.md - line 1086

* Update en.md - line 1128

* Update en.md  - line 1267

* Update en.md - line 1351

* Update en.md - line 1418

* Update en.md - line 1584

* Update en.md - line 1697

* Update en.md - line 1986

* Update en.md - line 2288

* Update en.md - line 2404

* Update en.md - line 2497

* Update en.md - line 2606

* Update en.md - line 2739

* Update en.md - line 2806

* Update en.md - line 2913

* Update en.md - line 3114

* Fix grammar and enhance clarity in en.md

Corrected grammatical errors and improved clarity in the text.

* Refine language and improve clarity in en.md

Revised text for clarity and conciseness throughout the document, including sections on fiat currency, historical financial events, and the relationship between money and inflation.

* Fix typos and enhance clarity in banking history text

Corrected typos and improved clarity in the text regarding the history of banking and currency.

* Refine language and punctuation in en.md

Corrected phrases for clarity and consistency, including changes to the naming of banks and punctuation adjustments throughout the document.

* Edit historical context for clarity and consistency

Revised text for clarity and consistency in the historical overview of reserve currencies and central banks.

* Refine text for clarity and conciseness in en.md

* Fix grammatical and punctuation errors in course content

Corrected punctuation and grammatical errors throughout the course description, enhancing clarity and readability.

* Fix grammatical errors and enhance text clarity

Corrected various grammatical errors and improved clarity throughout the text.

* Refine terminology and correct historical references

Updated various terms for clarity and consistency throughout the text, including corrections to the names of historical figures and concepts.

* Fix typographical errors and improve terminology consistency - final checks

Corrected minor typographical errors and improved consistency in terminology throughout the document, including the use of 'currencies' and 'lump'.

* Replace 'intermediary' with 'medium' in text

* Fix typo in sound money explanation

* general fixes

---------

Co-authored-by: MarJJ <marjjhodl@proton.me>